Modify compilation approach and add rustc compile error parsing.
Previously the compilation commands were executed in the project build
dir.
But rustc seems to have an issue with that because the compilation
errors are given with relative paths only.
So, considering that all the other backends have an option to specify
the compilation target dir from command line, then I forced all the
commands to execute form root now and specify the target from there.
With this approach all the current backends work properly and rust
errors in compilation buffer work as expected (can be clicked).